+++ tests/testthat/test-stackedDigitChart.R
@@ -2,16 +2,16 @@ test_that("stackedDigitChart produces an expected chart", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
-    onesDigit = sample(0:9, 1000, replace = TRUE)
+    onesPlace = sample(0:9, 1000, replace = TRUE)
   )
-  test_result <- stackedDigitChart(sample_data, siteID, onesDigit)
+  test_result <- stackedDigitChart(sample_data, siteID, onesPlace)
   expect_s3_class(test_result, "ggplot")
   expect_identical(as.character(test_result$labels$x), "siteID")
   expect_identical(as.character(test_result$labels$y), "Frequency")
   expect_identical(as.character(test_result$labels$fill), "Digit")
   expect_identical(
     test_result$labels$title,
-    "Digit Distribution of Data by siteID"
+    "Digit Distribution for onesPlace of Data by siteID"
   )
   vdiffr::expect_doppelganger(
     title = "stackedDigitChart default",
@@ -23,19 +23,38 @@ test_that("stackedDigitChart accepts supplied group label", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
-    onesDigit = sample(0:9, 1000, replace = TRUE)
+    onesPlace = sample(0:9, 1000, replace = TRUE)
   )
   test_result <- stackedDigitChart(
     sample_data,
     siteID,
-    onesDigit,
-    group_label = "Alternative Name"
+    onesPlace,
+    strGroupLabel = "Alternative Name"
   )
   expect_s3_class(test_result, "ggplot")
   expect_identical(test_result$labels$x, "Alternative Name")
   expect_identical(
     test_result$labels$title,
-    "Digit Distribution of Data by Alternative Name"
+    "Digit Distribution for onesPlace of Data by Alternative Name"
+  )
+})
+
+test_that("stackedDigitChart accepts supplied place label", {
+  set.seed(42)
+  sample_data <- data.frame(
+    siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
+    onesPlace = sample(0:9, 1000, replace = TRUE)
+  )
+  test_result <- stackedDigitChart(
+    sample_data,
+    siteID,
+    onesPlace,
+    strPlaceLabel = "Ones Place"
+  )
+  expect_s3_class(test_result, "ggplot")
+  expect_identical(
+    test_result$labels$title,
+    "Digit Distribution for Ones Place of Data by siteID"
   )
 })
 
@@ -43,14 +62,14 @@ test_that("stackedDigitChart accepts title override", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
-    onesDigit = sample(0:9, 1000, replace = TRUE)
+    onesPlace = sample(0:9, 1000, replace = TRUE)
   )
   test_result <- stackedDigitChart(
     sample_data,
     siteID,
-    onesDigit,
-    group_label = "Alternative Name",
-    chart_title = "My Chart"
+    onesPlace,
+    strGroupLabel = "Alternative Name",
+    strChartTitle = "My Chart"
   )
   expect_s3_class(test_result, "ggplot")
   expect_identical(test_result$labels$x, "Alternative Name")
@@ -60,19 +79,19 @@ test_that("stackedDigitChart accepts title override", {
   )
 })
 
-test_that("stackedDigitChart accepts digit_palette override", {
+test_that("stackedDigitChart accepts scaleDigitPalette override", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
-    onesDigit = sample(0:9, 1000, replace = TRUE)
+    onesPlace = sample(0:9, 1000, replace = TRUE)
   )
   vdiffr::expect_doppelganger(
     title = "stackedDigitChart no palette",
     fig = stackedDigitChart(
       sample_data,
       siteID,
-      onesDigit,
-      digit_palette = NULL
+      onesPlace,
+      scaleDigitPalette = NULL
     )
   )
   vdiffr::expect_doppelganger(
@@ -80,8 +99,8 @@ test_that("stackedDigitChart accepts digit_palette override", {
     fig = stackedDigitChart(
       sample_data,
       siteID,
-      onesDigit,
-      digit_palette = ggplot2::scale_fill_brewer(palette = "Paired")
+      onesPlace,
+      scaleDigitPalette = ggplot2::scale_fill_brewer(palette = "Paired")
     )
   )
 })
@@ -90,15 +109,15 @@ test_that("stackedDigitChart accepts theme override", {
   set.seed(42)
   sample_data <- data.frame(
     siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
-    onesDigit = sample(0:9, 1000, replace = TRUE)
+    onesPlace = sample(0:9, 1000, replace = TRUE)
   )
   vdiffr::expect_doppelganger(
     title = "stackedDigitChart no theme",
     fig = stackedDigitChart(
       sample_data,
       siteID,
-      onesDigit,
-      plot_theme = NULL
+      onesPlace,
+      themePlot = NULL
     )
   )
   vdiffr::expect_doppelganger(
@@ -106,8 +125,8 @@ test_that("stackedDigitChart accepts theme override", {
     fig = stackedDigitChart(
       sample_data,
       siteID,
-      onesDigit,
-      plot_theme = ggplot2::theme_minimal()
+      onesPlace,
+      themePlot = ggplot2::theme_minimal()
     )
   )
 })