+++ NEWS.md
@@ -1,3 +1,8 @@
 # gsm.digitpref (development version)
 
 - Added `stackedDigitChart()` for visualizing digit distribution by group. (@swanggilead, #5)
+- Added `splitByPlace()` to split a number into digits. (@jonthegeek, #6)
+
+# gsm.digitpref 0.0.0
+
+- Dummy version to prevent check warning before release.